refactor: split providers/mod.rs into per-type modules
- The file was 1825 lines against the 800 hard max, and the Provider
  serde change lands next. Moving it first keeps that diff out of a
  wall of relocated lines.
- Tests move to the file they cover, which is what forced the edgar
  injection and capability-bit cases out of the shared module.
- Move only: the same 1586 lib tests run before and after, and the
  crate-root exports are untouched.
